moved motion info into its own file
