+++ crates/gpui/src/gpui.rs
@@ -32,6 +32,7 @@ mod interactive;
 mod key_dispatch;
 mod keymap;
 mod lerp;
+mod motion_info;
 mod path_builder;
 mod platform;
 pub mod prelude;
@@ -141,6 +142,7 @@ pub use interactive::*;
 use key_dispatch::*;
 pub use keymap::*;
 pub use lerp::*;
+pub use motion_info::*;
 pub use path_builder::*;
 pub use platform::*;
 pub use profiler::*;
