+++ crates/gpui/src/style_transitions.rs
@@ -1,73 +1,6 @@
-use std::{
-    rc::Rc,
-    time::{Duration, Instant},
-};
+use std::time::Instant;
 
-use crate::Lerp;
-
-/// Converts a motion duration into a [`Duration`].
-pub trait IntoMotionDuration {
-    /// Converts this value into a duration.
-    fn into_motion_duration(self) -> Duration;
-}
-
-impl IntoMotionDuration for Duration {
-    fn into_motion_duration(self) -> Duration {
-        self
-    }
-}
-
-impl IntoMotionDuration for f32 {
-    fn into_motion_duration(self) -> Duration {
-        assert!(
-            self.is_finite() && self >= 0.0,
-            "motion duration must be a finite, non-negative number of seconds"
-        );
-        Duration::from_secs_f32(self)
-    }
-}
-
-/// Timing and easing configuration for a declarative style transition.
-#[derive(Clone)]
-pub struct MotionInfo {
-    duration: Duration,
-    easing: Rc<dyn Fn(f32) -> f32>,
-}
-
-impl MotionInfo {
-    /// Creates a linear motion with the supplied duration.
-    ///
-    /// A `Duration` is used as-is. An `f32` is interpreted as seconds.
-    pub fn new(duration: impl IntoMotionDuration) -> Self {
-        Self {
-            duration: duration.into_motion_duration(),
-            easing: Rc::new(crate::linear),
-        }
-    }
-
-    /// Replaces the linear easing function.
-    pub fn with_easing(mut self, easing: impl Fn(f32) -> f32 + 'static) -> Self {
-        self.easing = Rc::new(easing);
-        self
-    }
-
-    /// Returns the duration of this motion.
-    pub fn duration(&self) -> Duration {
-        self.duration
-    }
-}
-
-impl From<Duration> for MotionInfo {
-    fn from(duration: Duration) -> Self {
-        Self::new(duration)
-    }
-}
-
-impl From<f32> for MotionInfo {
-    fn from(seconds: f32) -> Self {
-        Self::new(seconds)
-    }
-}
+use crate::{Lerp, MotionInfo};
 
 pub(crate) struct StyleTransitionPropertyState<T> {
     goal_last_updated_at: Option<Instant>,
@@ -123,7 +56,7 @@ where
         self.start_goal = current_value;
         self.end_goal = goal;
 
-        if motion.duration.is_zero()
+        if motion.duration().is_zero()
             || self.start_goal.is_none()
             || self.end_goal.is_none()
             || self.start_goal == self.end_goal
@@ -141,7 +74,7 @@ where
             return (false, self.end_goal.clone());
         };
 
-        let duration = motion.duration.as_secs_f32();
+        let duration = motion.duration().as_secs_f32();
         if duration == 0.0 {
             self.jump_to(self.end_goal.clone());
             return (false, self.end_goal.clone());
@@ -180,15 +113,8 @@ gpui_macros::style_transitions!();
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        AbsoluteLength, DefiniteLength, InteractiveElement as _, Length, Style, div, ease_in_out,
-        px,
-    };
-
-    #[test]
-    fn f32_motion_duration_is_seconds() {
-        assert!((MotionInfo::new(0.2).duration().as_secs_f32() - 0.2).abs() < f32::EPSILON);
-    }
+    use crate::{AbsoluteLength, DefiniteLength, Length, Style, px};
+    use std::time::Duration;
 
     #[test]
     fn interrupted_property_motion_continues_from_its_current_value() {
@@ -240,30 +166,6 @@ mod tests {
             state.evaluate(Some(20.0), &motion, now, true),
             (false, Some(20.0))
         );
-    }
-
-    #[test]
-    fn generated_api_accepts_the_documented_motion_forms() {
-        let transitions = StyleTransitions::new()
-            .w(MotionInfo::new(Duration::from_millis(200)).with_easing(ease_in_out))
-            .bg(0.5)
-            .border_color(MotionInfo::new(0.6));
-
-        assert_eq!(
-            transitions.w.unwrap().duration(),
-            Duration::from_millis(200)
-        );
-        assert!((transitions.bg.unwrap().duration().as_secs_f32() - 0.5).abs() < f32::EPSILON);
-        assert!(
-            (transitions.border_color.unwrap().duration().as_secs_f32() - 0.6).abs() < f32::EPSILON
-        );
-
-        let mut built = false;
-        let _ = div().transitions(|transitions| {
-            built = true;
-            transitions.opacity(0.1)
-        });
-        assert!(built);
     }
 
     #[test]
